Do some project restructuring and cleanup.
