+++ source/SATextField/SATextField.h
@@ -1,0 +1,19 @@
+//
+//  SATextField.h
+//  SATextField
+//
+//  Created by Stefan Arambasich on 6/3/18.
+//  Copyright © 2018 Stefan Arambasich. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for SATextField.
+FOUNDATION_EXPORT double SATextFieldVersionNumber;
+
+//! Project version string for SATextField.
+FOUNDATION_EXPORT const unsigned char SATextFieldVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <SATextField/PublicHeader.h>
+
+
