+++ source/SATextFieldDemo/Images.xcassets/AppIcon.appiconset/Contents.json
