+++ source/SATextFieldTests/SATextFieldTests.swift
@@ -2,33 +2,35 @@
 //  SATextFieldTests.swift
 //  SATextFieldTests
 //
-//  Created by Stefan Arambasich on 9/3/2015.
-//  Copyright (c) 2015 Stefan Arambasich. All rights reserved.
+//  Created by Stefan Arambasich on 6/3/18.
+//  Copyright © 2018 Stefan Arambasich. All rights reserved.
 //
 
-import UIKit
 import XCTest
+@testable import SATextField
 
 class SATextFieldTests: XCTestCase {
-
+    
     override func setUp() {
         super.setUp()
+        // Put setup code here. This method is called before the invocation of each test method in the class.
     }
-
+    
     override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
         super.tearDown()
     }
-
+    
     func testExample() {
         // This is an example of a functional test case.
-        XCTAssert(true, "Pass")
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
     }
-
+    
     func testPerformanceExample() {
         // This is an example of a performance test case.
-        self.measure() {
+        self.measure {
             // Put the code you want to measure the time of here.
         }
     }
-
+    
 }
